requirements.txt
requests
beautifulsoup4
fpdf
schedule
python-dotenv
pandas